feature(deleteSelectedElements): add `deleteSelectedElements` method
